included rules to calculate scores and determine top3 personalities, sending results to the frontend.

+++ backend/src/main/java/riasec/backend/model/interfaces/TestAttempt.java
@@ -1,11 +1,9 @@
 package riasec.backend.model.interfaces;
 
-import riasec.backend.model.interfaces.Test;
-
 import java.util.Date;
 
 public interface TestAttempt {
-    Integer getId();
+    Long getId();
     Date getDate();
     Object getResult();
     Test getTest();
